完成商铺后台设计

+++ src/main/java/com/masou/coupon/dao/Shop2Dao.java
@@ -19,7 +19,7 @@ public class Shop2Dao {
     private ShopMapper shopMapper;
 
 
-    public int deleteByPrimaryKey(Integer id) {
+    public int deleteByPrimaryKey(Long id) {
         return shopMapper.deleteByPrimaryKey(id);
     }
